[feat] #34 모든 학생 조회 - 유저 타입에 따라 반환 값 다르게 설정
